requirements.txt
dwave-system==1.4.0
flake8==3.8.4
numpy==1.19.5
pandas==1.1.5
pyqubo==1.0.10
pytest==6.2.2
pytest-bdd==4.0.2
pytest-cov==2.11.1
python-dotenv==0.15.0
scipy==1.5.4
seaborn==0.11.1